# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "User registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User login success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "rt",
            "required": true,
            "in": "query",
            "description": "Refresh token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh tokens success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "Refresh auth tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/product": {
      "get": {
        "operationId": "ProductController_getProducts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items limit on page ( from 1 to 50 )",
            "schema": {
              "default": 10,
              "example": 15,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Items page ( min 1 )",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Product name",
            "schema": {
              "example": "nike",
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Sort by ( name, price, discount )",
            "schema": {
              "example": "price",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort method ( ASC | DESC )",
            "schema": {
              "example": "ASC",
              "type": "string"
            }
          },
          {
            "name": "priceFrom",
            "required": false,
            "in": "query",
            "description": "Product price from ( min 0 )",
            "schema": {
              "example": 500,
              "type": "number"
            }
          },
          {
            "name": "priceTo",
            "required": false,
            "in": "query",
            "description": "Product price to ( min 1 )",
            "schema": {
              "example": 1000,
              "type": "number"
            }
          },
          {
            "name": "categoryName",
            "required": false,
            "in": "query",
            "description": "Product category name",
            "schema": {
              "example": "Shoes",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Product category id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products recived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "Get list of products",
        "tags": [
          "Product"
        ]
      },
      "post": {
        "operationId": "ProductController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product",
        "tags": [
          "Product"
        ]
      }
    },
    "/product/{id}": {
      "get": {
        "operationId": "ProductController_getOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product recived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetails"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "Get product by id",
        "tags": [
          "Product"
        ]
      },
      "delete": {
        "operationId": "ProductController_removeProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product removed"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove product",
        "tags": [
          "Product"
        ]
      }
    },
    "/category": {
      "post": {
        "operationId": "CategoryController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create category",
        "tags": [
          "Category"
        ]
      },
      "get": {
        "operationId": "CategoryController_getRootCategories",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items limit on page ( from 1 to 50 )",
            "schema": {
              "default": 10,
              "example": 15,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Items page ( min 1 )",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Root categories received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootCategoriesDetail"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "Get root categories",
        "tags": [
          "Category"
        ]
      }
    },
    "/category/{id}/children": {
      "get": {
        "operationId": "CategoryController_getChildrenByParentId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Parent category id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcategories received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDetails"
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "summary": "Get subcategories",
        "tags": [
          "Category"
        ]
      }
    },
    "/category/{id}": {
      "delete": {
        "operationId": "CategoryController_removeCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "all",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          },
          "5XX": {
            "description": "Server error"
          },
          "4XX": {
            "description": "Check response message"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove category by id",
        "tags": [
          "Category"
        ]
      }
    }
  },
  "info": {
    "title": "Fake Store API",
    "description": "A simple REST API that simulates an online store — provides fake products, categories, and users for testing, prototyping, and learning purposes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductImageEntity": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Image url"
          }
        },
        "required": [
          "uuid"
        ]
      },
      "CategoryDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Category id"
          },
          "name": {
            "type": "string",
            "description": "Category name"
          },
          "hasChildren": {
            "type": "boolean",
            "description": "Is category has children"
          },
          "hasProduct": {
            "type": "boolean",
            "description": "Is category has product"
          }
        },
        "required": [
          "uuid",
          "name"
        ]
      },
      "ProductDetails": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Product id"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "price": {
            "type": "string",
            "description": "Product price"
          },
          "discount": {
            "type": "string",
            "description": "Product discount"
          },
          "images": {
            "description": "Product images",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageEntity"
            }
          },
          "category": {
            "description": "Product category",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryDetails"
              }
            ]
          }
        },
        "required": [
          "uuid",
          "name",
          "price",
          "discount",
          "images",
          "category"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first name ( from 1 to 20 symbols )",
            "example": "Evgeny"
          },
          "lastName": {
            "type": "string",
            "description": "User last name ( from 1 to 20 symbols )",
            "example": "Smirnov"
          },
          "login": {
            "type": "string",
            "description": "User login ( from 3 to 24 symbols )",
            "example": "nodewalker"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "nodewalker@yandex.com"
          },
          "password": {
            "type": "string",
            "description": "User password ( from 8 to 32 symbols )",
            "example": "**********"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "login",
          "email",
          "password"
        ]
      },
      "TokensDetails": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "refresh token"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "User login ( from 3 to 24 symbols )",
            "example": "nodewalker"
          },
          "password": {
            "type": "string",
            "description": "User password ( from 8 to 32 symbols )",
            "example": "**********"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "PaginationDetails": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total items"
          },
          "page": {
            "type": "number",
            "description": "Current page"
          },
          "limit": {
            "type": "number",
            "description": "Current limit"
          },
          "totalPage": {
            "type": "number",
            "description": "Total page count"
          },
          "isLastPage": {
            "type": "boolean",
            "description": "is last page?"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPage",
          "isLastPage"
        ]
      },
      "ProductsListDetails": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Product list",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDetails"
            }
          },
          "pagintaion": {
            "description": "Paggination info",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationDetails"
              }
            ]
          }
        },
        "required": [
          "data",
          "pagintaion"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of product ( from 2 to 50 symbols )",
            "example": "New Balance 1906"
          },
          "price": {
            "type": "number",
            "description": "Price of product ( min 1 )",
            "example": 1000
          },
          "categoryId": {
            "type": "string",
            "description": "Category id"
          },
          "discount": {
            "type": "number",
            "description": "Discount of product ( from 0 to 100 )",
            "default": 0,
            "example": 25
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            },
            "description": "Product images",
            "minItems": 1,
            "maxItems": 3
          }
        },
        "required": [
          "name",
          "price",
          "categoryId",
          "discount",
          "images"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of category ( from 2 to 50 symbols )",
            "example": "Shoes"
          },
          "parentId": {
            "type": "string",
            "description": "Parent category id"
          }
        },
        "required": [
          "name"
        ]
      },
      "RootCategoriesDetail": {
        "type": "object",
        "properties": {
          "tree": {
            "description": "Root categories tree",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDetails"
            }
          },
          "pagination": {
            "description": "Paggination info",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationDetails"
              }
            ]
          }
        },
        "required": [
          "tree",
          "pagination"
        ]
      }
    }
  }
}